# backend/openapi/swagger.yaml
components:
  schemas:
...
    github_com_uptaris_uptaris_backend_internal_inventory.IncidentRow:
      properties:
        createdAt:
          type: string
        description:
          maxLength: 10000
          type: string
        id:
          example: "0"
          type: string
        monitorId:
          example: "0"
          type: string
        monitorName:
          type: string
        resolvedAt:
          type: string
        serverId:
          example: "0"
          type: string
        severity:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
        startedAt:
          type: string
        status:
          enum:
          - open
          - acknowledged
          - resolved
          type: string
        title:
          maxLength: 200
          type: string
        updatedAt:
          type: string
      required:
      - severity
      - startedAt
      - status
      - title
      type: object
...
    github_com_uptaris_uptaris_backend_internal_models.Monitor:
      properties:
        createdAt:
          type: string
        expectedHealth:
          maxLength: 200
          type: string
        id:
          example: "0"
          type: string
        intervalSeconds:
          maximum: 86400
          minimum: 1
          type: integer
        lastCheckedAt:
          type: string
        name:
          maxLength: 200
          type: string
        serverId:
          example: "0"
          type: string
        status:
          enum:
          - up
          - down
          - paused
          type: string
        target:
          maxLength: 2048
          type: string
        type:
          enum:
          - http
          - tcp
          - icmp
          type: string
        updatedAt:
          type: string
      required:
      - expectedHealth
      - name
      - status
      - target
      - type
      type: object
...
    github_com_uptaris_uptaris_backend_internal_request.MonitorInput:
      properties:
        expectedHealth:
          example: "200"
          maxLength: 200
          type: string
        intervalSeconds:
          example: 60
          maximum: 86400
          minimum: 1
          type: integer
        name:
          example: Public API
          maxLength: 200
          type: string
        status:
          enum:
          - up
          - down
          - paused
          example: up
          type: string
        target:
          example: https://api.example.com/health
          maxLength: 2048
          type: string
        type:
          enum:
          - http
          - tcp
          - icmp
          example: http
          type: string
      required:
      - expectedHealth
      - intervalSeconds
      - name
      - target
      - type
      type: object
    github_com_uptaris_uptaris_backend_internal_request.MonitorPatchInput:
      properties:
        expectedHealth:
          example: "200"
          maxLength: 200
          type: string
        intervalSeconds:
          example: 60
          maximum: 86400
          minimum: 1
          type: integer
        name:
          example: Public API
          maxLength: 200
          type: string
        status:
          enum:
          - up
          - down
          - paused
          example: up
          type: string
        target:
          example: https://api.example.com/health
          maxLength: 2048
          type: string
        type:
          enum:
          - http
          - tcp
          - icmp
          example: http
          type: string
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.ErrorResponse:
      properties:
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.IncidentOverviewResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_inventory.IncidentRow'
          type: array
          uniqueItems: false
        links:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Links'
        pagination:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Pagination'
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.Links:
      properties:
        next:
          type: string
        previous:
          type: string
        self:
          type: string
      type: object
    github_com_uptaris_uptaris_backend_internal_response.MonitorListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Monitor'
          type: array
          uniqueItems: false
        links:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Links'
        pagination:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Pagination'
      type: object
    github_com_uptaris_uptaris_backend_internal_response.MonitorResponse:
      properties:
        _links:
          additionalProperties:
            type: string
          type: object
        data:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Monitor'
      type: object
    github_com_uptaris_uptaris_backend_internal_response.Pagination:
      properties:
        page:
          example: 1
          type: integer
        pageSize:
          example: 20
          type: integer
        total:
          example: 5
          type: integer
        totalPages:
          example: 1
          type: integer
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  description: REST API for Uptaris infrastructure monitoring.
  title: Uptaris API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /incidents:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      - description: Page size
        in: query
        name: pageSize
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Status
        in: query
        name: status
        schema:
          enum:
          - open
          - acknowledged
          - resolved
          type: string
      - description: Severity
        in: query
        name: severity
        schema:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.IncidentOverviewResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_query: invalid query parameter'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: List accessible incidents across servers
      tags:
      - incidents
...
  /servers/{serverId}/monitors:
    get:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      - description: Results per page
        in: query
        name: pageSize
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Type
        in: query
        name: type
        schema:
          enum:
          - http
          - tcp
          - icmp
          type: string
      - description: Status
        in: query
        name: status
        schema:
          enum:
          - up
          - down
          - paused
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.MonitorListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id or invalid_query: invalid ID or query parameter'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: List server monitors
      tags:
      - monitors
    post:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.MonitorInput'
        description: Monitor fields
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Monitor'
          description: Created
          headers:
            Location:
              description: Created resource URI
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id or invalid_json: invalid ID or JSON body'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'forbidden or origin_not_allowed: role or origin not permitted'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'validation_failed: invalid request field'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Create server monitor
      tags:
      - monitors
  /servers/{serverId}/monitors/{monitorId}:
    delete:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id: invalid resource ID'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'forbidden or origin_not_allowed: role or origin not permitted'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Delete monitor and incidents
      tags:
      - monitors
    get:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.MonitorResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id: invalid resource ID'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Get monitor
      tags:
      - monitors
    patch:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.MonitorPatchInput'
        description: Fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Monitor'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id or invalid_json: invalid ID or JSON body'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'forbidden or origin_not_allowed: role or origin not permitted'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'validation_failed: invalid request field'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Update monitor
      tags:
      - monitors
...
servers:
- url: /api/v1